feat: fix wrangler tests
